strip out validation into own file
