+++ eq-author-api/schema/resolvers/utils/index.js
@@ -1,5 +1,6 @@
 const getters = require("./getters");
 const helpers = require("./helpers");
 const theme = require("./theme");
+const validation = require("./validation");
 
-module.exports = { ...getters, ...helpers, ...theme };
+module.exports = { ...getters, ...helpers, ...theme, ...validation };
